# server/Src/Api/openapi.json
{
  "x-generator": "NSwag v14.6.3.0 (NJsonSchema v11.5.2.0 (Newtonsoft.Json v13.0.0.0))",
  "openapi": "3.0.0",
  "info": {
    "title": "Swagger UI",
    "version": "1.0.0"
  },
  "paths": {
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "Auth_Login",
        "requestBody": {
          "x-name": "loginRequest",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginCommand"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/register": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "Auth_Register",
        "requestBody": {
          "x-name": "registerRequest",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterUserCommand"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "operationId": "Auth_GetMe",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "Auth_Logout",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": []
          }
        ]
      }
    },
    "/api/realtime/GetTelemetry": {
      "get": {
        "tags": [
          "Realtime"
        ],
        "operationId": "Realtime_GetTelemetry",
        "parameters": [
          {
            "name": "connectionId",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RealtimeListenResponseOfListOfTelemetry"
                }
              }
            }
          }
        }
      }
    },
    "/api/realtime/sse": {
      "get": {
        "tags": [
          "Realtime"
        ],
        "operationId": "Realtime_Connect",
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "LoginCommand": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "RegisterUserCommand": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/RoleType"
          }
        }
      },
      "RoleType": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "User",
          "Crew",
          "Admin"
        ],
        "enum": [
          0,
          1,
          2
        ]
      },
      "RealtimeListenResponseOfListOfTelemetry": {
        "allOf": [
          {
            "$ref": "#/components/schemas/RealtimeListenResponse"
          },
          {
            "type": "object",
            "description": "Returned by subscribe endpoints with initial data. The client receives the current state\nimmediately and knows which SSE group to listen on for subsequent updates.",
            "additionalProperties": false,
            "properties": {
              "data": {
                "type": "array",
                "nullable": true,
                "items": {
                  "$ref": "#/components/schemas/Telemetry"
                }
              }
            }
          }
        ]
      },
      "Telemetry": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "guid"
          },
          "farmInternalId": {
            "type": "string",
            "format": "guid"
          },
          "farmId": {
            "type": "string"
          },
          "turbineInternalId": {
            "type": "string",
            "format": "guid"
          },
          "turbineId": {
            "type": "string"
          },
          "turbineName": {
            "type": "string"
          },
          "timeStamp": {
            "type": "string",
            "format": "date-time"
          },
          "windSpeed": {
            "type": "number",
            "format": "double"
          },
          "windDirection": {
            "type": "number",
            "format": "double"
          },
          "ambientTemperature": {
            "type": "number",
            "format": "double"
          },
          "rotorSpeed": {
            "type": "number",
            "format": "double"
          },
          "powerOutput": {
            "type": "number",
            "format": "double"
          },
          "nacelleDirection": {
            "type": "number",
            "format": "double"
          },
          "bladePitch": {
            "type": "number",
            "format": "double"
          },
          "generatorTemp": {
            "type": "number",
            "format": "double"
          },
          "gearboxTemp": {
            "type": "number",
            "format": "double"
          },
          "vibration": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "RealtimeListenResponse": {
        "type": "object",
        "description": "Returned by subscribe endpoints so the client knows which SSE group to listen on.",
        "additionalProperties": false,
        "properties": {
          "group": {
            "type": "string"
          }
        }
      }
    },
    "securitySchemes": {
      "JWT": {
        "type": "apiKey",
        "description": "JWT Authorization header using the Bearer scheme.",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}